requirements.txt
pandas
numpy
seaborn
scipy
matplotlib
hvplot
scikit-learn
xgboost
lightgbm
catboost
tensorflow-gpu==2.10.1
keras-tuner
pyyaml
-e .